+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Fixed forge particle registration (sery cutie)
 - Fixed issue when thigh high style list is empty (ThorTuwy)
 - Fixed Armor Trims not working on boobs (Doreena)
+- Gender change potion usable on Pokemon from the Cobblemon mod + estrogen effect